+++ back/main_sever/src/users/users.controller.ts
@@ -25,8 +25,8 @@ import { TempUserUseCase } from './usecase/temp-user.case';
 import { CommonExceptionFilter } from 'src/_common/filter/common-exception.filter';
 import { AccessTokenGuard } from 'src/auth/guard/token/access-token.guard';
 import { User } from './decorator/user.decorator';
-import { TicketUseCase } from 'src/tickets/usecase/ticket_use_case';
-import { TicketModel } from 'src/tickets/entities/ticket.entity';
+import { TicketUseCase } from 'src/users/_tickets/usecase/ticket_use_case';
+import { TicketModel } from 'src/users/_tickets/entities/ticket.entity';
 
 @Controller('users')
 export class UsersController {
