+++ back/main_sever/src/users/users.service.ts
@@ -8,7 +8,7 @@ import { UserModel } from './entities/user.entity';
 import { COMMON_FIND_USER_OPTIONS } from './const/user-find-options';
 import { DEFAULT_CREATE_USER_OPTION } from './const/default-user-create-option';
 import { TempUserModel } from './entities/temp-user.entity';
-import { TicketModel } from 'src/tickets/entities/ticket.entity';
+import { TicketModel } from 'src/users/_tickets/entities/ticket.entity';
 
 @Injectable()
 export class UsersService {
